pyproject.toml
[project]
name = "ultralytics-code-assistant"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.32.0",
    "pymongo>=4.6.0",
    "sentence-transformers>=2.5.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "indexer", "retriever", "generator", "config"]
